requirements.txt
fastapi
uvicorn
scikit-learn
joblib
nltk
python-multipart
mysql-connector-python